# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Aries Next API
  description: Aries Next 博客系统 API（由 utoipa 注解自动生成，权威契约即本文件）
  license:
    name: MIT
    identifier: MIT
  version: 0.1.0
servers:
- url: /
paths:
...
  /api/admin/categories:
    get:
      tags:
      - Admin Taxonomy
      summary: 获取文章分类列表
      description: 管理端分类接口固定 article kind（link / gallery 分类由各自模块管理），返回可用于 Article Editor 的 Category 列表。
      operationId: listCategories
      responses:
        '200':
          description: 分类列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 content:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    post:
      tags:
      - Admin Taxonomy
      summary: 新建文章分类
      description: 创建 article kind 分类；slug 缺省时以 name 归一化生成，唯一冲突返回 409 `TAXONOMY_CONFLICT`。
      operationId: createCategory
      requestBody:
        description: 分类名称与可选 Slug
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxonomyRequest'
        required: true
      responses:
        '201':
          description: 返回新建的分类
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 content:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: name 或 slug 已存在（`TAXONOMY_CONFLICT`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
  /api/admin/categories/{id}:
    put:
      tags:
      - Admin Taxonomy
      summary: 更新文章分类
      description: 更新分类的 `name` / `slug`；slug 复用统一 Normalize 规则，唯一冲突返回 409 `TAXONOMY_CONFLICT`。
      operationId: updateCategory
      parameters:
      - name: id
        in: path
        description: 分类 ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: 分类名称与可选 Slug
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxonomyRequest'
        required: true
      responses:
        '200':
          description: 返回更新后的分类
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 content:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 分类不存在（`TAXONOMY_NOT_FOUND`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: name 或 slug 已存在（`TAXONOMY_CONFLICT`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    delete:
      tags:
      - Admin Taxonomy
      summary: 删除文章分类
      description: 仍被 Article 引用时返回 409，并在 `error.details` 携带引用信息。
      operationId: deleteCategory
      parameters:
      - name: id
        in: path
        description: 分类 ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: 分类已删除
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 content:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 分类不存在（`TAXONOMY_NOT_FOUND`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 分类仍被引用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
...
  /api/admin/tags:
    get:
      tags:
      - Admin Taxonomy
      summary: 获取标签列表
      description: 返回可用于 Article Editor 的 Tag 列表。
      operationId: listTags
      responses:
        '200':
          description: 标签列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TagResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 content:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    post:
      tags:
      - Admin Taxonomy
      summary: 新建标签
      description: slug 缺省时以 name 归一化生成，唯一冲突返回 409 `TAXONOMY_CONFLICT`。
      operationId: createTag
      requestBody:
        description: 标签名称与可选 Slug
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxonomyRequest'
        required: true
      responses:
        '201':
          description: 返回新建的标签
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 content:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: name 或 slug 已存在（`TAXONOMY_CONFLICT`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
  /api/admin/tags/{id}:
    put:
      tags:
      - Admin Taxonomy
      summary: 更新标签
      description: 更新标签的 `name` / `slug`；唯一冲突返回 409 `TAXONOMY_CONFLICT`。
      operationId: updateTag
      parameters:
      - name: id
        in: path
        description: 标签 ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: 标签名称与可选 Slug
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaxonomyRequest'
        required: true
      responses:
        '200':
          description: 返回更新后的标签
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagResponse'
        '400':
          description: 参数校验失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 content:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 标签不存在（`TAXONOMY_NOT_FOUND`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: name 或 slug 已存在（`TAXONOMY_CONFLICT`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
    delete:
      tags:
      - Admin Taxonomy
      summary: 删除标签
      description: 仍被 Article 引用时返回 409，并在 `error.details` 携带引用信息。
      operationId: deleteTag
      parameters:
      - name: id
        in: path
        description: 标签 ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: 标签已删除
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无 content:manage 权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 标签不存在（`TAXONOMY_NOT_FOUND`）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 标签仍被引用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - cookieAuth: []
...
components:
  schemas:
...
    CategoryResponse:
      type: object
      title: CategoryResponse
      description: 管理端分类 DTO；`/api/admin/categories` 固定 `kind = article`，link / gallery 分类由各自模块管理。
      required:
      - id
      - kind
      - name
      - slug
      - description
      properties:
        description:
          type: string
        id:
          type: integer
          format: int64
        kind:
          type: string
          description: 分类类型（本模块固定为 `article`）。
        name:
          type: string
        parent_id:
          type:
          - integer
          - 'null'
          format: int64
        slug:
          type: string
...
    CreateTaxonomyRequest:
      type: object
      title: CreateTaxonomyRequest
      description: 创建/更新分类或标签的请求体。
      required:
      - name
      properties:
        name:
          type: string
          description: 显示名称；`slug` 缺省时以 name 归一化生成。
        slug:
          type:
          - string
          - 'null'
          description: 自定义 Slug；缺省时取 name。
...
    ErrorBody:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: 机器可读错误码，如 `UNAUTHORIZED`、`ARTICLE_NOT_FOUND`。
        details:
          description: 结构化补充信息（如引用计数），大多数错误为空，序列化时省略。
        message:
          type: string
          description: 人类可读错误信息。
    ErrorResponse:
      type: object
      description: |-
        统一错误响应体（与 `http::error::ApiError` 的实际 JSON 形状一致）：
        `{ "error": { "code", "message", "details?" } }`。
        所有端点的 4xx/5xx 响应统一引用本类型。
      required:
      - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
...
    TagResponse:
      type: object
      title: TagResponse
      description: 管理端标签 DTO。
      required:
      - id
      - name
      - slug
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: aries_admin_session